Shrink the lossless token stream and stop duplicating it
The lossless encoder allocates one token per pixel, so elosslessToken's size
is a per-pixel memory cost. Its five int fields came to 40 bytes; sizing each
to the VP8L limit it actually holds brings it to 16.

elosslessApplyColorCacheToTokens also built a second full token stream per
color-cache candidate while the base stream stayed live. It now writes into a
caller-supplied buffer: the cache-bits search reuses one scratch across
candidates, and the final rewrite runs in place over a base stream that is
dead by then.

Peak RSS on the 5.5 MP test image drops from 1299 to 923 MiB, and across the
photo corpus from 152-269 to 128-176 MiB per megapixel. Output is
byte-identical on every test image.

+++ enc_lossless_tokens.go
@@ -472,7 +472,7 @@ func elosslessBuildTokensGreedy(width int, argb []uint32, options elosslessToken
 		if bestMatch.set {
 			distance := bestMatch.distance
 			length := bestMatch.length
-			tokens = append(tokens, elosslessToken{kind: elosslessTokCopy, distance: distance, length: length})
+			tokens = append(tokens, elosslessToken{kind: elosslessTokCopy, distance: int32(distance), length: uint16(length)})
 			if cache != nil {
 				for _, pixel := range argb[index : index+length] {
 					cache.insert(pixel)
@@ -483,7 +483,7 @@ func elosslessBuildTokensGreedy(width int, argb []uint32, options elosslessToken
 			}
 			index += length
 		} else if cacheHit {
-			tokens = append(tokens, elosslessToken{kind: elosslessTokCache, key: cacheKey})
+			tokens = append(tokens, elosslessToken{kind: elosslessTokCache, key: uint16(cacheKey)})
 			if cache != nil {
 				cache.insert(argb[index])
 			}
@@ -871,10 +871,10 @@ func elosslessBuildTokensWithTraceback(width int, argb []uint32, options elossle
 			tokens = append(tokens, elosslessToken{kind: elosslessTokLiteral, argb: argb[cursor-1]})
 			cursor = previous[cursor]
 		case elosslessStepCache:
-			tokens = append(tokens, elosslessToken{kind: elosslessTokCache, key: st.key})
+			tokens = append(tokens, elosslessToken{kind: elosslessTokCache, key: uint16(st.key)})
 			cursor = previous[cursor]
 		case elosslessStepCopy:
-			tokens = append(tokens, elosslessToken{kind: elosslessTokCopy, distance: st.distance, length: st.length})
+			tokens = append(tokens, elosslessToken{kind: elosslessTokCopy, distance: int32(st.distance), length: uint16(st.length)})
 			start := cursor - st.length
 			if start < 0 {
 				start = 0
